fix: replaces Int with Long

+++ src/main/kotlin/com/erbanc/hunt/stats/bean/UserStatsBean.kt
@@ -4,11 +4,11 @@ class UserStatsBean(
 
     var username: String,
 
-    var questionsSolved: Int,
+    var questionsSolved: Long,
 
-    var totalBadGuesses: Int,
+    var totalBadGuesses: Long,
 
-    var totalGuesses: Int,
+    var totalGuesses: Long,
 
-    var rank: Int
+    var rank: Long
 )
